daily-news-summarizer/requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0

# HTTP and API clients
requests>=2.31.0
httpx>=0.27.0

# Environment and configuration
python-dotenv>=1.0.0

# Data handling
pandas

# Email
secure-smtplib>=0.1.1

# Async support
aiohttp>=3.9.0

# HTML parsing (for potential web scraping)
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Type checking and code quality
mypy>=1.7.0
black>=23.11.0
flake8>=6.1.0

# Logging
colorlog>=6.8.0
